Update README.md
had to add extended extension. You will need to add a temp folder and in that folder you will need to add a blank file called jigga.mp4 and the video popups will use that as a temp file to write to. I tried to upload folder to github but this time around, there is no way to add a folder, or upload it, even in chrome.